requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.9
djangorestframework==3.15.2
et-xmlfile==1.1.0
greenlet==3.1.1
kombu==5.4.2
mssql-django==1.5
mysqlclient==2.2.4
numpy==2.1.1
openpyxl==3.1.5
pandas==2.2.3
prompt_toolkit==3.0.48
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.35
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
